edith/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.47.0
django-cors-headers==3.11.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
idna==3.3
oauthlib==3.1.1
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
